From: Sudhir Kumar <skumar@linux.vnet.ibm.com>
libcgroup: testcases for wrapper api and cgroup_modify_cgroup()

This patch creates a valid cgroup structure to be passed as the parameter
for cgroup_modify_cgroup() function and calls this function. The return
values are checked and on that basis the test is declared as pass or fail.
The patch adds the code that check if the parameter changed reflect in
the filesystem.

Signed-off-by: Sudhir Kumar <skumar@linux.vnet.ibm.com>
Signed-off-by: Dhaval Giani <dhaval@linux.vnet.ibm.com>



git-svn-id: https://libcg.svn.sourceforge.net/svnroot/libcg/trunk@92 4f4bb910-9a46-0410-90c8-c897d4f1cd53