libcgroup: Fix index values in cgroup_add_value_*
There is a horrible bug in wrapper.c. The loop was causing a segmentation
fault. This is because we were checking for index value as opposed to
i.

This bug was not hit before, because in all our testing, we were
trying out only with CONFIG_FAIR_GROUP_SCHED and we were setting only one
value inside the memory subsystem.

Signed-off-by: Dhaval Giani <dhaval@linux.vnet.ibm.com>



git-svn-id: https://libcg.svn.sourceforge.net/svnroot/libcg/trunk@121 4f4bb910-9a46-0410-90c8-c897d4f1cd53