libcgroup: Add default group to cgconfig
From: Balbir Singh <balbir@linux.vnet.ibm.com>

As discussed in the previous thread
http://article.gmane.org/gmane.comp.lib.libcg.devel/139, we don't have
the concept of default groups and tasks should be able to return back
to their default group when done.

Here is a fix to the cgconfig script to fix the issue. Please review,
comment on it.

Signed-off-by: Balbir Singh <balbir@linux.vnet.ibm.com>
Signed-off-by: Dhaval Giani <dhaval@linux.vnet.ibm.com>



git-svn-id: https://libcg.svn.sourceforge.net/svnroot/libcg/trunk@238 4f4bb910-9a46-0410-90c8-c897d4f1cd53