libcgroup Test: libcgroup-modify_cgroup-test1-for-multimnt
This patch adds a testcase for libcgroup api cgroup_modify_cgroup() for
multiple mount scenario for first controller.

The challenge is to handle the control value as diff controllers put diff
upper limit on control value and we pick the controller dynamicaly.
We are going to use a config file for that which will have the controller name,
control file name and the control value. But not sure if there is any better
approach.
So thoughts ???

Signed-off-by: Sudhir Kumar <skumar@linux.vnet.ibm.com>

git-svn-id: https://libcg.svn.sourceforge.net/svnroot/libcg/trunk@262 4f4bb910-9a46-0410-90c8-c897d4f1cd53