libcgroup: Don't fail cgroup_get_cgroup if fscanf fails
cgroup_get_cgroup was failing with the memory controller enabled. It was
failing on memory.force_empty. The file has read permissions, but there 
is no read routine associated with it inside the kernel. fscanf failed  
and so cgroup_get_cgroup also failed. This was unexpected,              
cgroup_get_cgroup should have just skipped the file. In order to fix    
this, change cg_rd_ctrl_file to return more descriptive errors.         

Signed-off-by: Dhaval Giani <dhaval@linux.vnet.ibm.com>
Acked-by: Balbir Singh <balbir@linux.vnet.ibm.com>



git-svn-id: https://libcg.svn.sourceforge.net/svnroot/libcg/trunk@293 4f4bb910-9a46-0410-90c8-c897d4f1cd53