libcgrouptest: Change the dynamic creation of messages to static 2d array
From: Sudhir Kumar <skumar@linux.vnet.ibm.com>

This is addressing all the comments on the list and others too.

Earlier we were creating the messgae array dynamicaly which is not required
as the messages do not change dynamicaly. Let us have a static 2D array of
messages. This patch does that.
(This addresses the comment on the list ;))

Signed-off-by: Sudhir Kumar <skumar@linux.vnet.ibm.com>
Signed-off-by: Dhaval Giani <dhaval@linux.vnet.ibm.com>



git-svn-id: https://libcg.svn.sourceforge.net/svnroot/libcg/trunk@295 4f4bb910-9a46-0410-90c8-c897d4f1cd53