libcgrouptest: Move functions into a new file to reuse
From: Sudhir Kumar <skumar@linux.vnet.ibm.com>

There were some coding style warnings that have been fixed in this
patch. Basicaly the ones reported by checkpatch.pl


This patch puts the test functions in a separate file so that they
can be reused for further testing. Some of the functions were static
and hence changed to be non static. It also edits the Makefile so
that the patch compiles.

Signed-off-by: Sudhir Kumar <skumar@linux.vnet.ibm.com>
Signed-off-by: Dhaval Giani <dhaval@linux.vnet.ibm.com>



git-svn-id: https://libcg.svn.sourceforge.net/svnroot/libcg/trunk@309 4f4bb910-9a46-0410-90c8-c897d4f1cd53