From: Ivana Varekova <varekova@redhat.com>
This patch changes cgconfig.conf - comment all rows and rename
example values - to have example which should be installed to /etc/.
Move cgred to /etc/sysconfig and patch spec to install all 
/etc/* configuration files (at least they shows examples and 
help to user to do their own configuration).).

Signed-off-by: Ivana Varekova <varekova@redhat.com>
Signed-off-by: Balbir Singh <balbir@linux.vnet.ibm.com>

NOTE: The spec file is still broken due to lack of packaging of pam_cgroup.so.




git-svn-id: https://libcg.svn.sourceforge.net/svnroot/libcg/trunk@337 4f4bb910-9a46-0410-90c8-c897d4f1cd53