Add cgroup_get_procname_from_procfs() for getting a process name.
Hi,

Changelog of v6:
================
 * Change the returning values of *_get_procname_from_proc*() to integer
   from charactor pointer.
 * Clarify the number meaning of string length in cg_get_procname_from_
   ~proc_status()

Changelog of v5:
================
 * Rebase the patch to the latest code.

Changelog of v4:
================
 * Add the error handling for strdup()'s error.
 * Reduce strlen() calls.
 * Make the check code of a process name simple.

Changelog of v3:
================
 * Move cgroup_get_procname_from_procfs() to libcgroup-internal.h.
 * Fix unclear buffer of buf_cwd by memset().
 * Get a real path of script file by realpath().

Changelog of v2:
================
 * It is possible to handle a process, which name length is over than
   16 characters, also.

Description:
============
This patch adds a new function cgroup_get_procname_from_procfs()
for getting a process name.

This function allocates the memory for a process name, and writes
the name to the memory, and returns the pointer of the memory.
So a caller should free the memory if unusing it.

The process name, which is wrotten by this function, depends on
the specified process:

 If a command process)      the full path of command.
 If a shell script process) the full path of shell script.
 If a kernel thread)        the process name of kernel thread.

Thanks
Ken'ichi Ohmichi

Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>
Signed-off-by: Dhaval Giani <dhaval@linux.vnet.ibm.com>