Add the parser of process name in /etc/cgrules.conf.
Hi,

Changelog of v6:
================
 * The definations of CGROUP_RULE_MAXKEY and CGROUP_RULE_MAXLINE are
   moved to libcgroup-internal.h since no one from outside should be
   using them.

Changelog of v5:
================
 * Rebase the patch to the latest code.

Changelog of v4:
================
 * Use more safety length of a user name for the buffer "username".
 * Move the macros min()/max() to src/libcgroup-internal.h for using
   in src/api.c also.

Changelog of v3:
================
 * Fix unclear buffer of user by memset().

Changelog of v2:
================
 * Remove unnecessary memset().
 * Some cleanups.

Description:
============
This patch adds the parser of process name in /etc/cgrules.conf.

A new rule based on process name is as the following, and the process
name is stored into the member "procname" in struct cgroup_rule.
  <user>:<process name>  <controllers>   <destination>

Thanks
Ken'ichi Ohmichi

Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>
Signed-off-by: Dhaval Giani <dhaval@linux.vnet.ibm.com>