Don't link with -lpam unless necessary II
Oops, I sent older working version, here is correct one. I need to rest :(.
Here is the correct one:

All binaries are compiled with '-lpam' if --enable-pam is specified (and it
is on by default). We want only the PAM module linked with it.

The autoconf manual says about AC_CHECK_LIB:
	If action-if-found is not specified, the default action prepends
	-llibrary to LIBS

And the action-if-found we used was '[]', i.e. empty -> -lpam was added to
LIBS -> everything links with it. Simple action-if-found overrides this
default.

Changelog:
 - use 'dnl' as comments and ':' as the 'noop' action to reflect autoconf
 coding style.

Signed-off-by: Jan Safranek <jsafrane@redhat.com>
Signed-off-by: Balbir Singh <balbir@linux.vnet.ibm.com>