Fix a cgconfig behavior of wrong /etc/cgconfig.conf.
Changelog since v2:
o Avoid breaking ABI by changing the ECGCONFIGPARSEFAIL value.

Changelog since v1:
o Change a parser error value of /etc/cgconfig.conf to ECGCONFIGPARSEFAIL.

The cgconfig service succeeds even if there are something wrong
in /etc/cgconfig.conf. For example, it succeeds even if specifying
not-existing user (nouser) as a perm.task.uid like the following:

 # cat /etc/cgconfig.conf
 mount {
        cpuset = /mnt/cgroups/cpuset;
        memory = /mnt/cgroups/memory;
 }
 group root {
        perm {
                task {
                        uid = nouser;
                        gid = root;
                }
                admin {
                        uid = root;
                        gid = root;
                }
        }
        cpuset {
                cpuset.cpus = 0;
                cpuset.mems = 0;
        }
        memory {
                memory.use_hierarchy = 1;
                memory.limit_in_bytes = 1000000000000;
        }
 }
 #
 # service cgconfig start
 Starting cgconfig service: parsing failed at line number 10
                                                           [  OK  ]
 # echo $?
 0
 #

That is not a good behavior and the cgconfig service should fail
if there are something wrong in /etc/cgconfig.conf.
This patch fixes the behavior. In the above case, the behavior is
changed like the following by this patch:

 # service cgconfig start
 Starting cgconfig service: parsing failed at line number 10
 Loading configuration file /etc/cgconfig.conf failed
 Cgroup parsing failed
 Failed to parse /etc/cgconfig.conf                         [FAILED]
 #

Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>
Signed-off-by: Dhaval Giani <dhaval@linux.vnet.ibm.com>