Don't create sysdefault group whent it's not needed
IMHO most of users don't need sysdefault group being created during boot -
it's quite slow and at least in Fedora there is no application using it.

So, let users to configure it and allow them to turn the sysdefault off.

Whether the option should be enabled/disabled is distro-specific, we as
libcgroup turn it on.

Changelog:
 - create the sysdefault group by default

Signed-off-by: Jan Safranek <jsafrane@redhat.com>