libcgroup: Do not call exit from within the main library
The autogenerated code fromt the lexer was calling exit on a
fatal error. Do not do that. Instead log the message.

While it is still not clear if this is the correct response
for YY_FATAL_ERROR, we need to gracefully handle errors, and
since this is a path which is very rarely hit, this is still
a low risk change and so we will merge it.

Signed-off-by: Dhaval Giani <dhaval.giani@gmail.com>