Fix overlaping src and dest buffers in sprintf()
libcgroup compiled with various optimization flags produce strange results
when sprintf's destination and source pointers overlap or are the same.

In addition, use strncpy() instead. This is more or less useless, because
cg_build_path() does not check sizes either, but just for my personal feeling
of correct C code...

Signed-off-by: Jan Safranek <jsafrane@redhat.com>
Signed-off-by: Dhaval Giani <dhaval.giani@gmail.com>