cgclear: Do not rmdir the mount point
This code is strictly not needed, and was originally only added only
because cgconfigparser can mkdir the mountpoint.

Doing it however caused a few bugs and had the possibility of
changing the system state. So, let's not do it.

Should also fix up https://bugzilla.redhat.com/show_bug.cgi?id=652583

Signed-off-by: Dhaval Giani <dhaval.giani@gmail.com>
Tested-by:  Ivana Hutarova Varekova <varekova@redhat.com>
Signed-off-by: Jan Safranek <jsafrane@redhat.com>